#128: split the VsmcpTools catch-all + rename junk-drawer files
- Split the 1039-line VsmcpTools.cs: move the debug/breakpoints/inspection/
  memory/dump/processes/counters/profiler tools (~500 lines, 47 tools) into a
  new VsmcpTools.Debugging.cs partial. Base drops to ~630 lines. Pure code move
  (same partial class); the #127 parity test confirms no tool was lost.
- Rename junk-drawer files to descriptive names:
  Shared/Dtos.cs -> MetaDtos.cs; VsmcpTools.Many.cs -> VsmcpTools.BatchMany.cs;
  RpcTarget.Many.cs -> RpcTarget.BatchMany.cs; RpcTarget.CppMore.cs ->
  RpcTarget.CppMisc.cs; RpcTarget.EditMethod.cs -> RpcTarget.EditRefactor.cs.

No behavior change. 83 + 7 unit tests green (incl tool<->RPC parity); Server +
VSIX build. Full 1:1:1 topic alignment across all three projects remains
incremental follow-up.
