+++ src/VSMCP.Server/VsmcpTools.BatchMany.cs
