+++ src/VSMCP.Shared/MetaDtos.cs
