+++ src/VSMCP.Vsix/RpcTarget.BatchMany.cs
