+++ src/VSMCP.Vsix/RpcTarget.CppMisc.cs
