+++ src/VSMCP.Vsix/RpcTarget.EditRefactor.cs
