+++ src/VSMCP.Vsix/VSMCP.Vsix.csproj
@@ -85,7 +85,7 @@
     <Compile Include="FocusHelper.cs" />
     <Compile Include="FollowMode.cs" />
     <Compile Include="RpcTarget.FileMove.cs" />
-    <Compile Include="RpcTarget.EditMethod.cs" />
+    <Compile Include="RpcTarget.EditRefactor.cs" />
     <Compile Include="HostActivity.cs" />
     <Compile Include="StatusBarReporter.cs" />
     <Compile Include="VsmcpToolWindow.cs" />
@@ -135,9 +135,9 @@
     <Compile Include="VsmcpLog.cs" />
     <Compile Include="RpcTarget.CppSemantic.cs" />
     <Compile Include="RpcTarget.CppEdit.cs" />
-    <Compile Include="RpcTarget.CppMore.cs" />
+    <Compile Include="RpcTarget.CppMisc.cs" />
     <Compile Include="RpcTarget.CppRefactor.cs" />
-    <Compile Include="RpcTarget.Many.cs" />
+    <Compile Include="RpcTarget.BatchMany.cs" />
   </ItemGroup>
 
   <ItemGroup>
